requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.8
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
pillow>=10.0.0
celery>=5.3.0
redis>=4.6.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
whitenoise>=6.5.0
django-environ>=0.10.0

joblib>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0